Moved genericjsonio from CoreFunctionality to CoreUtilities
