+++ CoreDataAccessView/src/au/gov/asd/tac/constellation/views/dataaccess/io/ParameterIOUtilities.java
@@ -15,7 +15,7 @@
  */
 package au.gov.asd.tac.constellation.views.dataaccess.io;
 
-import au.gov.asd.tac.constellation.functionality.genericjsonio.JsonIO;
+import au.gov.asd.tac.constellation.utilities.genericjsonio.JsonIO;
 import au.gov.asd.tac.constellation.graph.Graph;
 import au.gov.asd.tac.constellation.graph.ReadableGraph;
 import au.gov.asd.tac.constellation.graph.WritableGraph;
