+++ CoreUtilities/nbproject/project.xml
@@ -140,6 +140,7 @@
                 <package>au.gov.asd.tac.constellation.utilities.datastructure</package>
                 <package>au.gov.asd.tac.constellation.utilities.file</package>
                 <package>au.gov.asd.tac.constellation.utilities.font</package>
+                <package>au.gov.asd.tac.constellation.utilities.genericjsonio</package>
                 <package>au.gov.asd.tac.constellation.utilities.geospatial</package>
                 <package>au.gov.asd.tac.constellation.utilities.graphics</package>
                 <package>au.gov.asd.tac.constellation.utilities.gui</package>
